requirements.txt
facenet-pytorch >= 2.5.3
pandas >= 2.0.2
numpy >= 1.22
pyyaml >= 5.1
annoy >= 1.17.3